+++ ach/src/main/java/com/adyen/checkout/ach/internal/ui/ACHDirectDebitDelegate.kt
@@ -3,11 +3,13 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by onurk on 17/1/2023.
+ * Created by onurk on 16/2/2023.
  */
 
-package com.adyen.checkout.ach
+package com.adyen.checkout.ach.internal.ui
 
+import com.adyen.checkout.ach.internal.ui.model.ACHDirectDebitInputData
+import com.adyen.checkout.ach.internal.ui.model.ACHDirectDebitOutputData
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
 import com.adyen.checkout.components.model.payments.request.ACHDirectDebitPaymentMethod
@@ -18,7 +20,7 @@ import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import com.adyen.checkout.core.exception.CheckoutException
 import kotlinx.coroutines.flow.Flow
 
-interface ACHDirectDebitDelegate :
+internal interface ACHDirectDebitDelegate :
     PaymentComponentDelegate<PaymentComponentState<ACHDirectDebitPaymentMethod>>,
     ViewProvidingDelegate,
     ButtonDelegate,
