+++ ach/src/main/java/com/adyen/checkout/ach/internal/ui/model/ACHDirectDebitOutputData.kt
@@ -3,17 +3,17 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by onurk on 24/1/2023.
+ * Created by onurk on 16/2/2023.
  */
 
-package com.adyen.checkout.ach
+package com.adyen.checkout.ach.internal.ui.model
 
 import com.adyen.checkout.components.base.OutputData
 import com.adyen.checkout.components.ui.AddressFormUIState
 import com.adyen.checkout.components.ui.AddressOutputData
 import com.adyen.checkout.components.ui.FieldState
 
-data class ACHDirectDebitOutputData(
+internal data class ACHDirectDebitOutputData(
     var bankAccountNumber: FieldState<String>,
     var bankLocationId: FieldState<String>,
     var ownerName: FieldState<String>,
