+++ ach/src/main/java/com/adyen/checkout/ach/internal/ui/model/AddressFieldPolicyParams.kt
@@ -3,10 +3,10 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by onurk on 24/1/2023.
+ * Created by onurk on 16/2/2023.
  */
 
-package com.adyen.checkout.ach
+package com.adyen.checkout.ach.internal.ui.model
 
 import com.adyen.checkout.components.ui.AddressFieldPolicy
 import kotlinx.parcelize.Parcelize
