+++ ach/src/test/java/com/adyen/checkout/ach/internal/util/ACHDirectDebitValidationUtilsTest.kt
@@ -3,11 +3,12 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by onurk on 31/1/2023.
+ * Created by onurk on 16/2/2023.
  */
 
-package com.adyen.checkout.ach
+package com.adyen.checkout.ach.internal.util
 
+import com.adyen.checkout.ach.R
 import com.adyen.checkout.components.ui.Validation
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
